pyproject.toml
[project]
name = "finscreener-mcp"
version = "0.1.0"
description = "MCP server for Finscreener API - Access company, director, GST data from any AI agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
finscreener-mcp = "mcp_server.server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]
